Change the type of the calendarlinkurl option to "select-input". This is to cover the following case:
An instance of Radarr or Sonarr is added to Organizr. Since the communication between Organizr and these services should preferably be done through a non-public channel, addresses like: http://radarr:7878 or http://sonarr:8989 are entered in the "Multiple URLs" fields. These URLs are then used to generate the links that lead from the calendar to the respective service. But if Oganizr is reachable from the public internet and is also used through it, the generated URLs do not work. With the new form type it is then possible to enter a self chosen public address.