Fix JavaScript ReferenceError: formatDuration is not defined
The real issue was that the JavaScript code was calling formatDuration() which doesn't exist.
The JellyStat plugin has its own formatJellyStatDuration() function that should be used instead.

This fixes the 'Uncaught ReferenceError: formatDuration is not defined' error that was preventing
the JellyStat homepage widget from displaying data after it successfully loaded from the API.